Move the encoding stuff into the encoding module
