+++ utils/bitfield/src/encoding/writer.rs
